requirements.txt

agno>=0.1.0
groq>=0.9.0
google-genai>=0.2.0
uvicorn[standard]>=0.24.0
websockets>=12.0
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
fastapi>=0.104.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
